Rename job_wrapper module
