+++ docs/compatibility.md
@@ -3,7 +3,8 @@ This document describes compatibility of the versions of this plugin with the Ra
 
 | Plugin version | Compatible Racetrack version |
 |----------------|------------------------------|
-| 2.9.1          | `> 2.18.0`                   |
-| 2.9.0          | `> 2.17.0`                   |
-| 2.8.0          | `>= 2.15.0`                  |
 | 2.6.2          | `>= 2.15.0`                  |
+| 2.8.0          | `>= 2.15.0`                  |
+| 2.9.0          | `>= 2.18.0`                  |
+| 2.9.1          | `>= 2.19.0`                  |
+| 2.9.2          | `>= 2.19.0`                  |
