+++ python3-job-type/job-template.Dockerfile
@@ -20,7 +20,7 @@ RUN . /src/job-venv/bin/activate &&\
 COPY . /src/job/
 RUN chmod -R a+rw /src/job/
 
-CMD python -u -m job_wrapper run "{{ manifest.get_jobtype_extra().entrypoint_path }}" "{{ manifest.get_jobtype_extra().entrypoint_class }}" < /dev/null
+CMD python -u -m racetrack_job_wrapper run "{{ manifest.get_jobtype_extra().entrypoint_path }}" "{{ manifest.get_jobtype_extra().entrypoint_class }}" < /dev/null
 
 ENV JOB_NAME "{{ manifest.name }}"
 ENV JOB_VERSION "{{ manifest.version }}"
