+++ python3-job-type/plugin-manifest.yaml
@@ -1,4 +1,4 @@
 name: python3-job-type
-version: 2.9.1
+version: 2.9.2
 url: 'https://github.com/TheRacetrack/plugin-python-job-type'
 category: 'job-type'
