+++ python3-job-type/python_wrapper/Makefile
@@ -14,4 +14,4 @@ test:
 # Run local server wrapping sample model (outside docker)
 run:
 	JOB_NAME=adder JOB_VERSION=0.0.1 \
-	job_wrapper run tests/sample/adder_model.py
+	racetrack_job_wrapper run tests/sample/adder_model.py
