--- python3-job-type/python_wrapper/job_wrapper/__main__.py
@@ -1,4 +1,0 @@
-from job_wrapper.main import main
-
-if __name__ == '__main__':
-    main()
