+++ python3-job-type/python_wrapper/racetrack_job_wrapper/__main__.py
@@ -1,0 +1,4 @@
+from racetrack_job_wrapper.main import main
+
+if __name__ == '__main__':
+    main()
