+++ python3-job-type/python_wrapper/racetrack_job_wrapper/loader.py
@@ -6,7 +6,7 @@ from importlib.abc import Loader
 from pathlib import Path
 from typing import Type, Optional
 
-from job_wrapper.entrypoint import JobEntrypoint
+from racetrack_job_wrapper.entrypoint import JobEntrypoint
 from racetrack_client.log.logs import get_logger
 
 logger = get_logger(__name__)
