+++ python3-job-type/python_wrapper/racetrack_job_wrapper/server.py
@@ -1,9 +1,9 @@
 import threading
 from typing import Optional
 
-from job_wrapper.api import create_health_app
-from job_wrapper.health import HealthState
-from job_wrapper.wrapper import create_entrypoint_app
+from racetrack_job_wrapper.api import create_health_app
+from racetrack_job_wrapper.health import HealthState
+from racetrack_job_wrapper.wrapper import create_entrypoint_app
 from racetrack_client.log.exception import short_exception_details
 from racetrack_client.log.logs import get_logger
 from racetrack_commons.api.asgi.asgi_reloader import ASGIReloader
