+++ python3-job-type/python_wrapper/racetrack_job_wrapper/validate.py
@@ -1,5 +1,5 @@
-from job_wrapper.docs import get_input_example
-from job_wrapper.entrypoint import JobEntrypoint, list_auxiliary_endpoints
+from racetrack_job_wrapper.docs import get_input_example
+from racetrack_job_wrapper.entrypoint import JobEntrypoint, list_auxiliary_endpoints
 from racetrack_client.log.context_error import wrap_context
 from racetrack_commons.api.response import ResponseJSONEncoder
