+++ python3-job-type/python_wrapper/tests/wrap/test_errors.py
@@ -1,6 +1,6 @@
-from job_wrapper.api import create_api_app
-from job_wrapper.entrypoint import JobEntrypoint
-from job_wrapper.health import HealthState
+from racetrack_job_wrapper.api import create_api_app
+from racetrack_job_wrapper.entrypoint import JobEntrypoint
+from racetrack_job_wrapper.health import HealthState
 from fastapi.testclient import TestClient
