+++ python3-job-type/python_wrapper/tests/wrap/test_perform_endpoint.py
@@ -1,4 +1,4 @@
-from job_wrapper.wrapper import create_entrypoint_app
+from racetrack_job_wrapper.wrapper import create_entrypoint_app
 from fastapi.testclient import TestClient
