+++ python3-job-type/python_wrapper/tests/wrap/test_webview.py
@@ -6,7 +6,7 @@ from fastapi.testclient import TestClient
 import pytest
 from starlette.routing import Mount
 
-from job_wrapper.wrapper import create_entrypoint_app
+from racetrack_job_wrapper.wrapper import create_entrypoint_app
 
 
 @pytest.fixture(scope="function")
